Working "copy CSS" button added.
Renamed all instances of HTML element "copyButton" to "copyButtonHTML"

Added a new section in script.j starting on row 210. The logic for the "copy CSS" button is placed in this section.

Added an extra "Copy CSS" button on every card. Shifted bottom margin (where needed) to that button instead. The buttons are looped out

Added an array in script.js that holds the current paths to the card css files and the ID:s for the "Copy CSS" button that correlates with that path. More ID:s and paths can be added as new card css files are added when new Profile card templates arrives.

Created the asynchronous function "getCssFileContents", that fetches css file contents as text.